requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.3.0
click==8.1.3
colorama==0.4.6
Django==4.2.2
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.3
PyJWT==2.7.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
